machine-learning-client/requirements.txt
opencv-python
numpy
torch
torchvision
facenet_pytorch
onnxruntime-gpu